requirements.txt
python-dotenv
streamlit
pandas
lightgbm
numpy
openpyxl
scikit-learn==1.6.1
supabase